Moved CallbackData to its own file
And change its visibility to pub(crate).
This required turning functions in evntrace to pub(crate) as well.
